BAVL-284 refactoring of emails. Pulling out court and probation emails into their own separate files.
